fix syntax

+++ Program.cs
@@ -8,9 +8,9 @@ class MainClass
         Console.Write("Calculadora de hipotenusa!");
         Console.WriteLine();
         Console.Write("digite um valor para o primeiro cateto: ");
-        cateto1 = int.Parse(Console.ReadLine());
+        cateto1 = double.Parse(Console.ReadLine());
         Console.Write("digite um valor para o segundo cateto: ");
-        cateto2 = int.Parse(Console.ReadLine());
+        cateto2 = double.Parse(Console.ReadLine());
 
         hipotenusa = Math.Sqrt(Math.Pow(cateto1, 2) + Math.Pow(cateto2, 2));
